# docs/02-springboot-foundation/openapi.yaml
openapi: 3.0.3

info:
  title: Posts API
  description: REST API for managing posts
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Health
    description: Application health endpoints

  - name: Posts
    description: Operations for managing posts

paths:

  /api/health:

    get:
      tags:
        - Health

      summary: Check API health

      description: Returns the current application health status.

      operationId: getHealth

      responses:
        '200':
          description: Application is running
          content:
            text/plain:
              schema:
                type: string
                example: Posts API is running!


  /api/posts:

    get:
      tags:
        - Posts

      summary: Get all posts

      description: Returns all posts.

      operationId: getAllPosts

      responses:
        '200':
          description: Posts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'


    post:
      tags:
        - Posts

      summary: Create a post

      description: Creates a new post.

      operationId: createPost

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostRequest'

      responses:

        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'

        '400':
          description: Invalid request


  /api/posts/{id}:

    parameters:

      - name: id
        in: path
        required: true

        description: Unique identifier of the post

        schema:
          type: integer
          format: int64

          example: 1


    get:
      tags:
        - Posts

      summary: Get a post

      description: Returns a post by its ID.

      operationId: getPostById

      responses:

        '200':
          description: Post found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'

        '404':
          description: Post not found


    put:
      tags:
        - Posts

      summary: Update a post

      description: Updates an existing post.

      operationId: updatePost

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostRequest'

      responses:

        '200':
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'

        '404':
          description: Post not found


    delete:
      tags:
        - Posts

      summary: Delete a post

      description: Deletes an existing post.

      operationId: deletePost

      responses:

        '204':
          description: Post deleted successfully

        '404':
          description: Post not found


components:

  schemas:

    Post:
      type: object

      description: A post resource.

      properties:

        id:
          type: integer
          format: int64
          readOnly: true
          example: 1

        title:
          type: string
          example: First Post

        body:
          type: string
          example: Hello Spring Boot


    PostRequest:
      type: object

      required:
        - title
        - body

      properties:

        title:
          type: string
          example: First Post

        body:
          type: string
          example: Hello Spring Boot